+++ kanban/release-notes.md
@@ -1,6 +1,10 @@
 #Release Notes
 
-## 0.0.3
+## 0.2.0
+[X] Retrieves content of the topic Object Orientation
+[X] Removes undesired 'tab' character in final of phrases in DB
+
+## 0.1.0
 [x] View database sheet content in front
 [x] Integration with Google Analytics